+++ extensions/api/data-management/asset/src/test/java/org/eclipse/dataspaceconnector/api/datamanagement/asset/service/AssetServiceImplTest.java
@@ -1,6 +1,5 @@
-package org.eclipse.dataspaceconnector.api.datamanagement.asset;
+package org.eclipse.dataspaceconnector.api.datamanagement.asset.service;
 
-import org.eclipse.dataspaceconnector.api.datamanagement.asset.service.AssetServiceImpl;
 import org.eclipse.dataspaceconnector.dataloading.AssetLoader;
 import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.spi.asset.AssetIndex;
@@ -42,7 +41,7 @@ class AssetServiceImplTest {
     void findById_shouldRelyOnAssetIndex() {
         when(index.findById("assetId")).thenReturn(createAsset("assetId"));
 
-        var asset = service.findbyId("assetId");
+        var asset = service.findById("assetId");
 
         String assetId = "assetId";
         assertThat(asset).isNotNull().matches(hasId(assetId));