+++ spi/core-spi/src/test/java/org/eclipse/dataspaceconnector/spi/query/QuerySpecTest.java
@@ -83,4 +83,5 @@ class QuerySpecTest {
 
         assertThat(qs.getFilterExpression()).isNotNull().isEmpty();
     }
+
 }