+++ CHANGELOG.md
@@ -29,10 +29,13 @@ the detailed section referring to by linking pull requests or issues.
 * Implement S3BucketReader (#675)
 * Add configuration setting for state machine batch size (#872)
 * Add Jetty context alias for IDS API (#815)
-* Implement Asset service for Data Management API (#931)
 * Add `Hostname` service (#948)
-* Added embedded and remote DPF Selector (#832)
+* Add embedded and remote DPF Selector (#832)
 * Pass path information into http data source through DPF public API (#929)
+* Add instructions for observability sample with Azure Application Insights (#928)
+* Add interface `WebServer` to `web-spi` (#921)
+* Implement Asset service for Data Management API (#931)
+* Implement ContractDefinition service for Data Management API (#940)
 
 #### Changed
 
@@ -47,6 +50,10 @@ the detailed section referring to by linking pull requests or issues.
 * Update build system to Java 17 (#934)
 * Refactor (=generify) transformer subsystem (#779)
 * Extract interfaces for every api controller class to improve swagger documentation (#891)
+* Instrument executors with metrics (#912)
+* Call the listeners before the state transition is persisted. (#876)
+* Added an overload to `TransactionContext#execute()` (#968)
+* Run CosmosDB integration tests on cloud in CI (#964)
 
 #### Removed